# If applied, this commit will...
# Why is this change needed?

# How does it address the issue?

# Provide links to any relevant tickets, articles or other resources
Fixed the average_vote table key. This key need to report a float to give the data more revelance.

+++ db/migrate/20191115221238_create_movies.rb
@@ -7,7 +7,7 @@ class CreateMovies < ActiveRecord::Migration[6.0]
       t.text :poster_path
       t.text :backdrop_path
       t.text :overview
-      t.integer :average_vote
+      t.float :average_vote
 
       t.timestamps null: false
     end
